+++ elements/wifi/sr/localbroadcast.hh
@@ -57,7 +57,7 @@ private:
   Timer _timer;
   IPAddress _ip;    // My IP address.
   EtherAddress _en; // My ethernet address.
-  uint16_t _et;     // This protocol's ethertype
+  uint32_t _et;     // This protocol's ethertype
   IPAddress _bcast_ip;
 
   EtherAddress _bcast;
